worked on routing&&favoritesPage
