+++ src/App.css
@@ -3,7 +3,6 @@
   width: 100vw;
   height: 100vh;
   box-sizing: border-box;
-  
 }
 
 .wrapper{
