+++ src/App.js
@@ -7,6 +7,7 @@ import SignUpPage from './pages/SignUpPage/SignUpPage';
 import SignInPage from './pages/SignInPage/SignInPage';
 import Home from './pages/Home/Home';
 import Header from './components/Header/Header';
+import Favorites from './pages/Favorites/Favorites';
 
 function App() {
   return (
@@ -16,6 +17,7 @@ function App() {
         <Header/>
           <Routes>
             <Route path = '/' element = {<Home/>}/>
+            <Route path ='/Favorites' element = {<Favorites/>}/>
             <Route path='/SignInPage' element = {<SignInPage/>}/>
             <Route path='/SignUpPage' element = {<SignUpPage/>}/>
           </Routes>
