+++ src/components/Card/Card.jsx
@@ -1,16 +1,24 @@
 import "./Card.css"
 import {useDispatch, useSelector} from 'react-redux'
-import { addFavoritesAction } from '../../store/Reducers/favoritesReducer'
-import React from "react";
+import { addFavoritesAction, removeFavoritesAction } from '../../store/Reducers/favoritesReducer'
+import React, { useState } from "react";
 
 const Card = ({card}) => {
+  const [fav,setFav] = useState(false)
   const dispatch = useDispatch()
   const favorites = useSelector(state=>state.favorites.userFavorites)
   const displayAuthor = card.authors[0]?.name || 'Anonymous'
 
   const addToFavFunc = () =>{
     dispatch(addFavoritesAction({id: card.id, title: card.title, author: displayAuthor}))
+    setFav(true)
+    
   }
+
+  // const removeFromFavFunc= () =>{
+  //   dispatch(removeFavoritesAction(card.id))
+  //   setFav(false)
+  // }
 
   
     return (
